Implement noindex & skip brackets (#112)
Implements bracket support for folder names. With a bracket you can skip
or ignore parts of the slug, see the following use-cases:

1. The `blog/[2025]/[01]/example-post` folder structure means the
`blog/example-post` slug. This way you don't need to add `noindex.yml`
files in the 2025, and 01 folders.

2. Use [] to hide extra stuff from the slug, e.g. `[01]example-post` ->
`example-post`. This is extremely useful when we have ordered content
and want to keep the directories in sync with the order.